Introduction

This package is currently under active development. The final goal is to have several modules.

National

State

Household

Bilateral Trade

GTAP

Currently, the only available module is the National module.

DataFrames.subset — Method

subset(
        data::T, 
        @nospecialize(args...)
    ) where T <: WiNDCtable

Return a subset of data based on the conditions given in args. This will subset both the main table and the set table.

Required Arguments

data::T: The WiNDCtable to subset.

args::Tuple: Pairs of the form set_name => boolean function.

Return

A table of type T.

WiNDC.aggregate — Method

function aggregate(
    data::T,
    aggregations...
) where T<:WiNDCtable

Required Arguments

data::T: The WiNDCtable to aggregate.

aggregations: Takes the form set_name => (X, original => new)).

set_name is a symbol, the name of the set to aggregate.

X a dataframe with the columns original and new.

original is the name of the column in with the elements to be aggregated.

new is the name of the column with the aggregated names.

WiNDC.calibrate — Method

calibrate(data::AbstractNationalTable)

This is currently geared toward calibrating the national dataset. I'll be working to make this be a general calibration function.

Returns a new AbstractNationalTable with the calibrated values and the model.

There are three primary balancing operations:

Zero Profit - Column sums are equal

Market Clearance - Row sums are equal

Margin Balance - The margins balance

The three tax rates are fixed. The tax rates are:

Output Tax Rate

Absorption Tax Rate

Import Tariff Rate

The following are fixed:

Labor Compensation

Imports

Exports

Household Supply

Any zero values will remain zero.

WiNDC.calibrate — Method

calibrate(data::AbstractNationalTable)

This is currently geared toward calibrating the national dataset. I'll be working to make this be a general calibration function.

Returns a new AbstractNationalTable with the calibrated values and the model.

There are three primary balancing operations:

Zero Profit - Column sums are equal

Market Clearance - Row sums are equal

Margin Balance - The margins balance

The three tax rates are fixed. The tax rates are:

Output Tax Rate

Absorption Tax Rate

Import Tariff Rate

The following are fixed:

Labor Compensation

Imports

Exports

Household Supply

Any zero values will remain zero.

WiNDC.create_national_sets — Method

create_national_sets(
    use::XLSX.Worksheet, 
    supply::XLSX.Worksheet,
    set_regions)

This function creates the sets for the detailed national data.

set regions for detailed table

Dict(
    "commodities" => ("use", ["A7:B408"], false, :commodities),
    "labor_demand" => ("use", ["A410:B410"], false, :commodities),
    "other_tax" => ("use", ["A411:B411"], false, :commodities),
    "capital_demand" => ("use", ["A412:B412"], false, :commodities),
    "sectors" => ("use", ["C5:ON6"], true, :sectors),
    "personal_consumption" => ("use", ["OP5:OP6"], true, :sectors),
    "household_supply" => ("use", ["OP5:OP6"], true, :sectors),
    "exports" => ("use", ["OV5:OV6"], true, :sectors),
    "exogenous_final_demand" => ("use", ["OQ5:OU6","OW5:PH6"], true, :sectors),
    "imports" => ("supply", ["OP5:OP6"], true, :sectors),
    "margin_demand" => ("supply", ["OS5:OT6"], true, :sectors),
    "margin_supply" => ("supply", ["OS5:OT6"], true, :sectors),
    "duty" => ("supply", ["OV5:OV6"], true, :sectors),
    "tax" => ("supply", ["OW5:OW6"], true, :sectors),
    "subsidies" => ("supply", ["OX5:OX6"], true, :sectors)
)

WiNDC.create_national_subtables — Method

create_national_subtables(sets)

This function creates the subtables for the detailed national data.

WiNDC.detailed_summary_map — Method

detailed_summary_map(detailed_path)

This function reads the detailed table and returns a DataFrame that maps the detailed sectors to the summary sectors. The first sheet of the detailed table is a map between the detailed sectors and the summary sectors. In addition this maps value added, final demand and supply extras to the summary sectors.

WiNDC.down_fill — Method

down_fill(X)

This function fills in the missing values in a column with the last non-missing value.

WiNDC.get_set — Method

get_set(data::T) where T<:WiNDCtable  

get_set(data::T, set_name::String) where T<:WiNDCtable

get_set(data::T, set_name::Vector{String}) where T<:WiNDCtable

Return the elements of the given sets. If no set is given, return all sets.

Required Arguments

data - A WiNDCtable-like object.

set_name - A string or vector of strings representing the set names to be extracted.

Returns

Returns a DataFrame with three columns, :element, :set and :description

WiNDC.get_subtable — Method

get_subtable(data::T, subtable::String, column::Vector{Symbol}; negative::Bool = false, keep_all_columns = false) where T<:WiNDCtable

get_subtable(data::T, subtable::String; column::Symbol = :value, output::Symbol = :value, negative = false) where T<:WiNDCtable

get_subtable(data::T, subtable::Vector{String}) where T<:WiNDCtable

Return the subtables requested as a DataFrame

Required Arguments

data - A WiNDCtable-like object.

subtable - A string or vector of strings representing the subtable names to be extracted.

Optional Arguments

column - A symbol representing the column to be extracted. Default is :value.

output - A symbol representing the output column name. Default is :value.

negative - A boolean representing whether the values should be negated. Default is false.

Returns

Returns a DataFrame with the requested subtables and columns.

WiNDC.get_table — Method

get_table(data::T) where T<:WiNDCtable

Return the main table of the WiNDCtable object as a DataFrame

Required Arguments

data - A WiNDCtable-like object.

Output

Returns a DataFrame with columns domain(data), subtable, and value.

WiNDC.load_table — Method

load_table(
    file_path::String
    years::Int...;
)

Load a WiNDCtable from a file.

Required Arguments

file_path::String: The path to the file.

years::Int...: The years to load. If no years are provided, all years in the file will be loaded.

Returns

A subtype of a WiNDCtable, with the data and sets loaded from the file.

WiNDC.make_subtable — Method

make_subtable(sets, rows, columns, table, subtable)

A helper function for extracting subtables.

WiNDC.save_table — Method

save_table(
    output_path::String
    MU::T,
) where T<:WiNDCtable

Save a WiNDCtable to a file. The file format is HDF5, which can be opened in any other language. The file will have the following structure:

year - DataFrame - The data for each year in the WiNDCtable sets - DataFrame - The sets of the WiNDCtable columns - Array - The column names of each yearly DataFrame

Required Arguments

output_path::String: The path to save the file.

MU::WiNDCtable: The WiNDCtable to save.

WiNDC.weight_function — Method

weight_function(year_detail, year_summary, minimum_detail, maximum_detail)

Create the weight function for the interpolation of the detailed table to the summary table based solely on the year.