National Disaggregation Model

WiNDC.national_model_mcp_year — Method

national_model_mcp_year(GU::GamsUniverse,year::Symbol;solver = PATHSolver.Optimizer)

Run the national model for a single year.

WiNDC.national_model_mcp — Method

national_model_mcp(GU::GamsUniverse;solver = PATHSolver.Optimizer)

Run all years of the the national model. Returns a dictionary containing each year.

To Do:

Create documentation explaining exactly what the model is doing

Currently this model is hard-coded to be the benchmark

verification. Need to make modifications to also run a counterfactual.