State Disaggregation Model

WiNDC.state_disaggregation_model_mcp_year — Method

state_disaggregation_model_mcp_year(GU::GamsUniverse,year::Symbol)

Return the model for the state-level disaggreation for a single year.

WiNDC.state_disaggregation_model_mcp — Method

state_disaggregation_model_mcp(GU::GamsUniverse)

Run all years of the the state-level model. Returns a dictionary containing each year.

To Do:

Create documentation explaining exactly what the model is doing

Currently this model is hard-coded to be the benchmark

verification. Need to make modifications to also run a counterfactual.