Introduction

Currently only the National model is implemented. In time the plan is to rewrite the entire WiNDC build stream in Julia.

Example

API Version

Currently (Sept 28, 2023) the API only gives data for years 2017 - 2022.

In order to use this you must have an api key from the BEA. Register here to obtain an key.

using WiNDC
using GamsStructure

api_key = #paste API key here
GU = load_bea_data_api(api_key;years = 1997:2021);

models = national_model_mcp(GU)

Local Version

This data is hosted on the WiNDC website. Download and extract the zip file windc_2021.zip. The tables are located in the directory:

BEA/IO

using WiNDC
using GamsStructure

GU = load_bea_data_local("path_to_use_table",
                         "path_to_supply_table",
);

models = national_model_mcp(GU)

Accessing Models

m = models[1997]

print(generate_report(m))

Function Listing

WiNDC.load_bea_data_api — Method

load_bea_data_api(api_key::String; years = 1997:2021)

Load the the BEA data using the BEA API.

In order to use this you must have an api key from the BEA. Register here to obtain an key.

Currently (Septerber 28, 2023) this will only return years 2017-2022 due to the BEA restricting the API.

WiNDC.load_bea_data_local — Method

load_bea_data_local(use_path::String,
                    supply_path::String;
                    years = 1997:2021)

Load the BEA data from a local XLSX file. This data is available at the WiNDC webpage. The use table is

windc_2021/BEA/IO/Use_SUT_Framework_1997-2021_SUM.xlsx

and the supply table is

windc_2021/BEA/IO/Supply_Tables_1997-2021_SUM.xlsx

WiNDC.national_model_mcp — Method

national_model_mcp(GU::GamsUniverse;solver = PATHSolver.Optimizer)

Run all years of the the national model. Returns a dictionary containing each year.